+++ src/lib/forms/Toggle.svelte
@@ -1,14 +1,23 @@
 <script lang="ts">
   import { getContext } from 'svelte';
+  import type { ComponentProps } from 'svelte';
   import { twMerge } from 'tailwind-merge';
   import Checkbox from './Checkbox.svelte';
   import type { FormColorType } from '../types';
 
-  export let size: 'small' | 'default' | 'large' | 'custom' = 'default';
-  export let group: (string | number)[] = [];
-  export let value: string | number = '';
-  export let checked: boolean | undefined = undefined;
-  export let customSize: string = '';
+  interface $$Props extends Omit<ComponentProps<Checkbox>, 'size'> {
+    size?: 'small' | 'default' | 'large' | 'custom';
+    group?:  string[];
+    value?: string | number;
+    checked?: boolean;
+    customSize?: string;
+  }
+
+  export let size: NonNullable<$$Props['size']> = 'default';
+  export let group: $$Props['group'] = [];
+  export let value: $$Props['value'] = '';
+  export let checked: $$Props['checked'] = undefined;
+  export let customSize: $$Props['customSize'] = '';
 
   // tinted if put in component having its own background
   let background: boolean = getContext('background');
@@ -49,9 +58,9 @@
 @component
 [Go to docs](https://flowbite-svelte.com/)
 ## Props
-@prop export let size: 'small' | 'default' | 'large' | 'custom' = 'default';
-@prop export let group: (string | number)[] = [];
-@prop export let value: string | number = '';
-@prop export let checked: boolean | undefined = undefined;
-@prop export let customSize: string = '';
+@prop export let size: NonNullable<$$Props['size']> = 'default';
+@prop export let group: $$Props['group'] = [];
+@prop export let value: $$Props['value'] = '';
+@prop export let checked: $$Props['checked'] = undefined;
+@prop export let customSize: $$Props['customSize'] = '';
 -->