+++ src/lib/types.ts
@@ -9,21 +9,11 @@ export type ButtonColorType = 'alternative' | 'blue' | 'dark' | 'green' | 'light
 
 export type CheckboxItem = {
   value: string;
-  label?: string; 
-  isChecked?: boolean; 
+  label?: string;
+  isChecked?: boolean;
 };
 
-export type ColorVariant =
-  | 'dark'
-  | 'red'
-  | 'yellow'
-  | 'green'
-  | 'indigo'
-  | 'purple'
-  | 'pink'
-  | 'blue'
-  | 'primary'
-  | 'none';
+export type ColorVariant = 'dark' | 'red' | 'yellow' | 'green' | 'indigo' | 'purple' | 'pink' | 'blue' | 'primary' | 'none';
 
 export type Colors = 'blue' | 'gray' | 'red' | 'yellow' | 'purple' | 'green' | 'indigo' | 'pink' | 'white' | 'custom' | 'primary' | 'secondary';
 
@@ -79,6 +69,7 @@ export type ReviewType = {
 export type SelectOptionType<T> = {
   name: string | number;
   value: T;
+  disabled?: boolean;
 };
 
 export type TransitionTypes = 'fade' | 'fly' | 'slide' | 'blur' | 'in:fly' | 'out:fly' | 'in:slide' | 'out:slide' | 'in:fade' | 'out:fade' | 'in:blur' | 'out:blur';
@@ -163,9 +154,4 @@ export interface ButtonClassesTypes {
   custom?: string;
 }
 
-export type ToastPositionType =
-  | 'top-left'
-  | 'top-right'
-  | 'bottom-left'
-  | 'bottom-right'
-  | 'none';
+export type ToastPositionType = 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | 'none';